requirements.txt
selenium>=4.15.0
spotipy>=2.23.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
